requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
filelock==3.12.2
h11==0.14.0
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.25.2
opencv-python-headless==4.8.0.76
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyclipper==1.3.0.post4
PyKakao==0.0.4
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
selenium==4.11.2
shapely==2.0.1
six==1.16.0
slack-sdk==3.21.3
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
tifffile==2023.7.18
torch==2.0.1
torchvision==0.15.2
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0
